Enable the updateRegistrarRdapBaseUrls endpoint
This endpoint will update the Registrar RDAP Base URLs from ICANN once a
day.

Note that the endpoint needs a TLD to use for the query, since the ICANN
MoSAPI endpoint we use needs a TLD to authenticate (even though the
results don't depend on the TLD)

This is unfortunate in an open source project, and we might need to
change it to a configured value.